server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Venu Thottempudi",email = "vthottempudi1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.125.0,<0.126.0)",
    "supabase (>=2.27.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyjwt (>=2.8.0,<3.0.0)",
    "clerk-backend-api (>=1.0.0,<2.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "boto3 (>=1.42.16,<2.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "langchain (==0.3.27)",
    "langchain-community (==0.3.27)",
    "langchain-openai (==0.3.28)",
    "unstructured[all-docs] (==0.18.11)",
    "scrapingbee (>=2.0.2,<3.0.0)"
]




[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
